Add files via upload
This lightweight JavaScript code allows you to create a rudimentary age calculator.  It calculates a person's age depending on the birth date supplied into a form.  When a user enters his or her birthday, the calculator fetches the current date and time, calculates the difference in days between the two dates, calculates the person's age in years, and presents the results in an HTML element.   In addition, the code performs the computations using the built-in JavaScript Date object and Math functions, and then updates the HTML element with the age value via the innerHTML property.